Removed MockSearch duplication
