+++ autofit/mock/mock.py
@@ -2,12 +2,30 @@ import math
 import numpy as np
 from typing import Optional
 
+from autoconf import conf
+
 from autofit.non_linear.analysis import Analysis
 from autofit.mapper.model import ModelInstance
 from autofit.mapper.model_mapper import ModelMapper
 from autofit.non_linear.abstract_search import NonLinearSearch
+from autofit.non_linear.result import Result
 from autofit.non_linear.samples import PDFSamples, Sample, NestSamples
-from autoconf import conf
+
+from autofit import exc
+
+
+def samples_with_log_likelihood_list(
+        log_likelihood_list
+):
+    return [
+        Sample(
+            log_likelihood=log_likelihood,
+            log_prior=0,
+            weight=0
+        )
+        for log_likelihood
+        in log_likelihood_list
+    ]
 
 
 class MockAnalysis(Analysis):
@@ -28,7 +46,18 @@ class MockAnalysis(Analysis):
         pass
 
 
-class MockResult:
+class MockAnalysis2(Analysis):
+    def log_likelihood_function(self, instance):
+        return 1.0
+
+    def visualize(self, paths, instance, during_analysis):
+        pass
+
+    def __init__(self, data):
+        self.data = data
+
+
+class MockResult(Result):
     def __init__(
             self,
             samples=None,
@@ -37,7 +66,10 @@ class MockResult:
             analysis=None,
             search=None,
     ):
-        self.instance = instance or ModelInstance()
+
+        super().__init__(samples, model, search)
+
+        self._instance = instance or ModelInstance()
         self.model = model or ModelMapper()
         self.samples = samples or MockSamples(max_log_likelihood_instance=self.instance)
 
@@ -88,7 +120,7 @@ class MockSamples(PDFSamples):
         if self._log_likelihood_list is not None:
             log_likelihood_list = self._log_likelihood_list
         else:
-            log_likelihood_list = range(5)
+            log_likelihood_list = range(3)
 
         return [
             Sample(
@@ -126,6 +158,9 @@ class MockSamples(PDFSamples):
             return super().gaussian_priors_at_sigma(sigma=sigma)
 
         return self._gaussian_tuples
+
+    def write_table(self, filename):
+        pass
 
 
 class MockNestSamples(NestSamples):
@@ -178,8 +213,20 @@ class MockNestSamples(NestSamples):
 
 
 class MockSearch(NonLinearSearch):
-    def __init__(self, samples=None, result=None, name="", prior_passer=None):
-        super().__init__(name=name)
+    def __init__(
+            self,
+            name="",
+            samples=None,
+            result=None,
+            unique_tag: Optional[str] = None,
+            prior_passer=None,
+            fit_fast=True,
+            sample_multiplier=1,
+            save_for_aggregator=False,
+            **kwargs
+    ):
+
+        super().__init__(name=name, unique_tag=unique_tag, **kwargs)
 
         self.samples = samples or MockSamples()
         self.result = result or MockResult(samples=samples)
@@ -187,7 +234,20 @@ class MockSearch(NonLinearSearch):
         if prior_passer is not None:
             self.prior_passer = prior_passer
 
-    def _fit(self, model, analysis, log_likelihood_cap=None):
+        self.fit_fast = fit_fast
+        self.sample_multiplier = sample_multiplier
+
+        self.save_for_aggregator = save_for_aggregator
+
+    @property
+    def config_type(self):
+        return conf.instance["non_linear"]["mock"]
+
+    @property
+    def config_dict_search(self):
+        return {}
+
+    def _fit_fast(self, model, analysis, log_likelihood_cap=None):
         class Fitness:
             def __init__(self, instance_from_vector, result):
                 self.result = result
@@ -204,23 +264,62 @@ class MockSearch(NonLinearSearch):
                 # Return Chi squared
                 return -2 * log_likelihood
 
-        analysis.save_attributes_for_aggregator(paths=self.paths)
+        if self.save_for_aggregator:
+            analysis.save_attributes_for_aggregator(paths=self.paths)
 
         fitness_function = Fitness(model.instance_from_vector, result=self.result)
-        fitness_function(model.prior_count * [0.5])
+        fitness_function(model.prior_count * [0.8])
 
         return fitness_function.result
 
-    @property
-    def config_type(self):
-        return conf.instance["non_linear"]["mock"]
+    def _fit(self, model, analysis, log_likelihood_cap=None):
+
+        if self.fit_fast:
+            result = self._fit_fast(model=model, analysis=analysis)
+            return result
+
+        if model.prior_count == 0:
+            raise AssertionError("There are no priors associated with the model!")
+        if model.prior_count != len(model.unique_prior_paths):
+            raise AssertionError(
+                "Prior count doesn't match number of unique prior paths"
+            )
+        index = 0
+        unit_vector = model.prior_count * [0.5]
+        while True:
+            try:
+                instance = model.instance_from_unit_vector(unit_vector)
+                fit = analysis.log_likelihood_function(instance)
+                break
+            except exc.FitException as e:
+                unit_vector[index] += 0.1
+                if unit_vector[index] >= 1:
+                    raise e
+                index = (index + 1) % model.prior_count
+        samples = MockSamples(
+            sample_list=samples_with_log_likelihood_list(self.sample_multiplier * fit),
+            model=model,
+            gaussian_tuples=[
+                (prior.mean, prior.width if math.isfinite(prior.width) else 1.0)
+                for prior in sorted(model.priors, key=lambda prior: prior.id)
+            ],
+        )
+
+        self.paths.save_samples(samples)
+
+        return MockResult(
+            model=model,
+            samples=samples,
+        )
 
     def perform_update(self, model, analysis, during_analysis):
-        self.paths.save_object("samples", self.samples)
-        return self.samples
-
-    def samples_from(self, model):
-        return self.samples
+        return MockSamples(
+            sample_list=samples_with_log_likelihood_list([1.0, 2.0]),
+            gaussian_tuples=[
+                (prior.mean, prior.width if math.isfinite(prior.width) else 1.0)
+                for prior in sorted(model.priors, key=lambda prior: prior.id)
+            ]
+        )
 
 
 class Profile:
