+++ test_autofit/config/non_linear/mock/MockSearchBC.ini
@@ -1,0 +1,19 @@
+[search]
+
+[updates]
+iterations_per_update=2500
+visualize_every_update=1
+model_results_every_update=1
+log_every_update=1
+remove_state_files_at_end=True
+
+[initialize]
+method=prior
+
+[printing]
+silence=False
+
+[prior_passer]
+sigma=3.0
+use_errors=True
+use_widths=True
