+++ test_autofit/graphical/gaussian/test_global_fit.py
@@ -6,6 +6,7 @@ import autofit as af
 import autofit.graphical as g
 from autofit.tools.namer import namer
 
+from autofit.mock.mock import MockSearch
 
 @pytest.fixture(
     autouse=True
@@ -104,7 +105,7 @@ class TestGlobalLikelihood:
 
     def test_global_search(self, model_factor, model_factor_2):
         collection = g.FactorGraphModel(model_factor, model_factor_2)
-        search = af.MockSearch()
+        search = MockSearch()
 
         class Analysis(af.Analysis):
             def log_likelihood_function(self, instance):
