+++ test_autofit/non_linear/grid/conftest.py
@@ -2,7 +2,7 @@ import pytest
 
 import autofit as af
 from autofit.mock.mock_model import MockClassx2Tuple
-
+from autofit.mock.mock import MockSearch
 
 @pytest.fixture(name="mapper")
 def make_mapper():
@@ -15,7 +15,7 @@ def make_mapper():
 
 @pytest.fixture(name="grid_search")
 def make_grid_search(mapper):
-    mock_search = af.MockSearch()
+    mock_search = MockSearch()
     mock_search.paths = af.DirectoryPaths(name="")
     search = af.SearchGridSearch(
         number_of_steps=10,
