+++ test_autofit/non_linear/grid/test_paths/test_indicators.py
@@ -5,7 +5,9 @@ import pytest
 import autofit as af
 from autoconf.conf import output_path_for_test
 from autofit.database.aggregator.scrape import Scraper
-from autofit.mock.mock import MockAnalysis
+
+from autofit.mock.mock import MockAnalysis, MockSearch
+
 from test_autofit.non_linear.grid.test_optimizer_grid_search import MockOptimizer
 
 output_directory = Path(
@@ -20,8 +22,8 @@ output_directory = Path(
     output_directory
 )
 def make_parent_search(model_gaussian_x1):
-    search = af.MockSearch(
-        "parent"
+    search = MockSearch(
+        name="parent"
     )
     search.paths.model = model_gaussian_x1
     return search
