+++ test_autofit/non_linear/grid/test_result_builder.py
@@ -6,6 +6,7 @@ from autofit.non_linear.grid.grid_search import ResultBuilder
 from autofit.non_linear.grid.grid_search.job import JobResult
 from autofit.non_linear.grid.grid_search.result_builder import Placeholder
 
+from autofit.mock.mock import MockSearch, MockResult
 
 @pytest.fixture(
     name="result_builder"
@@ -27,8 +28,8 @@ def make_add_results(
         for number in numbers:
             result_builder.add(
                 JobResult(
-                    af.MockResult(
-                        search=af.MockSearch(
+                    MockResult(
+                        search=MockSearch(
                             name=str(number)
                         ),
                         model=af.Model(
