+++ test_autofit/non_linear/test_result.py
@@ -2,7 +2,7 @@ import pytest
 import autofit as af
 
 from autofit.mock.mock_model import MockClassx2Tuple
-from autofit.mock.mock_search import MockSamples, MockSearch
+from autofit.mock.mock import MockSamples, MockSearch
 
 @pytest.fixture(name="result")
 def make_result():
@@ -41,7 +41,7 @@ class TestResult:
     def test_raises(self, result):
         with pytest.raises(af.exc.PriorException):
             result.model.mapper_from_gaussian_tuples(
-                result.samples.gaussian_tuples, a=2.0, r=1.0
+                result.samples._gaussian_tuples, a=2.0, r=1.0
             )
